backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
alembic
python-dotenv
python-multipart
httpx
pydantic[email]
